+++ slimta/util/proxyproto.py
@@ -28,13 +28,37 @@ services.
 
 from __future__ import absolute_import
 
+import struct
+
 from gevent import socket
 
 from slimta.logging import getSocketLogger
 
-__all__ = ['ProxyProtocolV1']
+__all__ = ['ProxyProtocol', 'ProxyProtocolV1', 'ProxyProtocolV2']
 
 log = getSocketLogger(__name__)
+
+#: The source address returned if UNKNOWN is given in the proxy protocol
+#: header.
+unknown_pp_source_address = (None, None)
+
+#: The destination address returned if UNKNOWN is given in the proxy
+#: protocol header.
+unknown_pp_dest_address = (None, None)
+
+#: The source address returned if there was a parsing error or EOF while
+#: reading the proxy protocol header.
+invalid_pp_source_address = (None, None)
+
+#: The destination address returned if there was a parsing error or EOF
+#: while reading the proxy protocol header.
+invalid_pp_dest_address = (None, None)
+
+
+class LocalConnection(Exception):
+    # Used to indicate that the parsed proxy protocol header is for a "local"
+    # connection, and should not be proxied.
+    pass
 
 
 class ProxyProtocolV1(object):
@@ -51,23 +75,8 @@ class ProxyProtocolV1(object):
 
     """
 
-    #: The source address returned if UNKNOWN is given in the proxy protocol
-    #: header.
-    unknown_pp_source_address = (None, None)
-
-    #: The destination address returned if UNKNOWN is given in the proxy
-    #: protocol header.
-    unknown_pp_dest_address = (None, None)
-
-    #: The source address returned if there was a parsing error or EOF while
-    #: reading the proxy protocol header.
-    invalid_pp_source_address = (None, None)
-
-    #: The destination address returned if there was a parsing error or EOF
-    #: while reading the proxy protocol header.
-    invalid_pp_dest_address = (None, None)
-
-    def __read_pp_line(self, sock, initial):
+    @classmethod
+    def __read_pp_line(cls, sock, initial):
         buf = bytearray(107)
         buf[0:len(initial)] = initial
         read = initial
@@ -86,7 +95,8 @@ class ProxyProtocolV1(object):
                 break
         return read
 
-    def parse_pp_line(self, line):
+    @classmethod
+    def parse_pp_line(cls, line):
         """Given a bytestring containing a single line ending in CRLF, parse
         into two source and destination address tuples of the form
         ``(ip, port)`` and return them.
@@ -102,17 +112,18 @@ class ProxyProtocolV1(object):
         line = line[6:-2]
         parts = line.split(b' ')
         if parts[0] == b'UNKNOWN':
-            return self.unknown_pp_source_address, self.unknown_pp_dest_address
-        family = self.__get_pp_family(parts[0])
+            return unknown_pp_source_address, unknown_pp_dest_address
+        family = cls.__get_pp_family(parts[0])
         assert len(parts) == 5, \
             'Invalid proxy protocol header format'
-        source_addr = (self.__get_pp_ip(family, parts[1], 'source'),
-                       self.__get_pp_port(parts[3], 'source'))
-        dest_addr = (self.__get_pp_ip(family, parts[2], 'destination'),
-                     self.__get_pp_port(parts[4], 'destination'))
+        source_addr = (cls.__get_pp_ip(family, parts[1], 'source'),
+                       cls.__get_pp_port(parts[3], 'source'))
+        dest_addr = (cls.__get_pp_ip(family, parts[2], 'destination'),
+                     cls.__get_pp_port(parts[4], 'destination'))
         return source_addr, dest_addr
 
-    def __get_pp_family(self, family_string):
+    @classmethod
+    def __get_pp_family(cls, family_string):
         if family_string == b'TCP4':
             return socket.AF_INET
         elif family_string == b'TCP6':
@@ -120,7 +131,8 @@ class ProxyProtocolV1(object):
         else:
             raise AssertionError('Invalid proxy protocol address family')
 
-    def __get_pp_ip(self, addr_family, ip_string, which):
+    @classmethod
+    def __get_pp_ip(cls, addr_family, ip_string, which):
         try:
             packed = socket.inet_pton(addr_family, ip_string.decode('ascii'))
             return socket.inet_ntop(addr_family, packed)
@@ -128,7 +140,8 @@ class ProxyProtocolV1(object):
             msg = 'Invalid proxy protocol {0} IP format'.format(which)
             raise AssertionError(msg)
 
-    def __get_pp_port(self, port_string, which):
+    @classmethod
+    def __get_pp_port(cls, port_string, which):
         try:
             port_num = int(port_string)
         except ValueError:
@@ -138,21 +151,168 @@ class ProxyProtocolV1(object):
             'Proxy protocol {0} port out of range'.format(which)
         return port_num
 
+    @classmethod
+    def process_pp_v1(cls, sock, initial):
+        line = cls.__read_pp_line(sock, initial)
+        log.recv(sock, line)
+        return cls.parse_pp_line(line)
+
     def handle(self, sock, addr):
         """Intercepts calls to :meth:`~slimta.edge.EdgeServer.handle`, reads
         the proxy protocol header, and then resumes the original call.
 
         """
         try:
-            line = self.__read_pp_line(sock, b'')
-            log.recv(sock, line)
-            src_addr, _ = self.parse_pp_line(line)
+            src_addr, _ = self.process_pp_v1(sock, b'')
         except AssertionError as exc:
             log.proxyproto_invalid(sock, exc)
-            src_addr = self.invalid_pp_source_address
+            src_addr = invalid_pp_source_address
         else:
             log.proxyproto_success(sock, src_addr)
         return super(ProxyProtocolV1, self).handle(sock, src_addr)
+
+
+class ProxyProtocolV2(object):
+    """Implements version 2 of the proxy protocol, to avoid losing information
+    about the original connection when routing traffic through a proxy. This
+    version is binary, and may be more efficient than version 1.
+
+    Mix-in before an implementation of :class:`~slimta.edge.EdgeServer` to
+    expect every connection to begin with a proxy protocol header. The
+    ``address`` argument passed in to the
+    :meth:`~slimta.edge.EdgeServer.handle` method will contain information
+    about the original connection source, before proxying.
+
+    """
+
+    __commands = {0x00: 'local',
+                  0x01: 'proxy'}
+    __families = {0x10: socket.AF_INET,
+                  0x20: socket.AF_INET6,
+                  0x30: socket.AF_UNIX}
+    __protocols = {0x01: socket.SOCK_STREAM,
+                   0x02: socket.SOCK_DGRAM}
+
+    @classmethod
+    def __read_pp_data(cls, sock, length, initial):
+        buf = bytearray(length)
+        buf[0:len(initial)] = initial
+        read = initial
+        while len(read) < len(buf):
+            where = memoryview(buf)[len(read):]
+            read_n = sock.recv_into(where, len(buf)-len(read))
+            assert read_n, 'Received EOF during proxy protocol header'
+            read = memoryview(buf)[0:len(read)+read_n].tobytes()
+        return bytearray(read)
+
+    @classmethod
+    def __parse_pp_data(cls, data):
+        assert data[0:12] == b'\r\n\r\n\x00\r\nQUIT\n', \
+            'Invalid proxy protocol v2 signature'
+        assert data[13] & 0xf0 == 0x20, 'Invalid proxy protocol version'
+        command = cls.__commands.get(data[12] & 0x0f)
+        family = cls.__families.get(data[13] & 0xf0)
+        protocol = cls.__protocols.get(data[13] & 0x0f)
+        addr_len = struct.unpack('<H', data[14:16])[0]
+        return command, family, protocol, addr_len
+
+    @classmethod
+    def __parse_pp_addresses(cls, family, addr_data):
+        if family == socket.AF_INET:
+            src_ip, dst_ip, src_port, dst_port = \
+                struct.unpack('<4s4sHH', addr_data)
+            src_addr = (socket.inet_ntop(family, src_ip), src_port)
+            dst_addr = (socket.inet_ntop(family, dst_ip), dst_port)
+            return src_addr, dst_addr
+        elif family == socket.AF_INET6:
+            src_ip, dst_ip, src_port, dst_port = \
+                struct.unpack('<16s16sHH', addr_data)
+            src_addr = (socket.inet_ntop(family, src_ip), src_port)
+            dst_addr = (socket.inet_ntop(family, dst_ip), dst_port)
+            return src_addr, dst_addr
+        elif family == socket.AF_UNIX:
+            src_addr, dst_addr = struct.unpack('<108s108s', addr_data)
+            return src_addr.rstrip(b'\x00'), dst_addr.rstrip(b'\x00')
+        else:
+            return unknown_pp_source_address,  unknown_pp_dest_address
+
+    @classmethod
+    def process_pp_v2(cls, sock, initial):
+        try:
+            data = cls.__read_pp_data(sock, 16, initial)
+            cmd, family, _, addr_len = cls.__parse_pp_data(data)
+            addr_data = cls.__read_pp_data(sock, addr_len, b'')
+            ret = cls.__parse_pp_addresses(family, addr_data)
+            if cmd == 'local':
+                raise LocalConnection()
+            return ret
+        except struct.error:
+            raise AssertionError('Invalid proxy protocol data')
+
+    def handle(self, sock, addr):
+        """Intercepts calls to :meth:`~slimta.edge.EdgeServer.handle`, reads
+        the proxy protocol header, and then resumes the original call.
+
+        """
+        try:
+            src_addr, _ = self.process_pp_v2(sock, b'')
+        except LocalConnection:
+            log.proxyproto_local(sock)
+            return
+        except AssertionError as exc:
+            log.proxyproto_invalid(sock, exc)
+            src_addr = invalid_pp_source_address
+        else:
+            log.proxyproto_success(sock, src_addr)
+        return super(ProxyProtocolV2, self).handle(sock, src_addr)
+
+
+class ProxyProtocol(object):
+    """Reads the first 8 bytes from the socket to determine which version of
+    the proxy protocol is active before handing processing off to either
+    :class:`ProxyProtocolV1` or :class:`ProxyProtocolV2`.
+
+    Mix-in before an implementation of :class:`~slimta.edge.EdgeServer` to
+    expect every connection to begin with a proxy protocol header. The
+    ``address`` argument passed in to the
+    :meth:`~slimta.edge.EdgeServer.handle` method will contain information
+    about the original connection source, before proxying.
+
+    """
+
+    @classmethod
+    def __read_pp_initial(cls, sock):
+        buf = bytearray(8)
+        read = b''
+        while len(read) < len(buf):
+            where = memoryview(buf)[len(read):]
+            read_n = sock.recv_into(where, 8-len(read))
+            assert read_n, 'Received EOF during proxy protocol header'
+            read = memoryview(buf)[0:len(read)+read_n].tobytes()
+        return read
+
+    def handle(self, sock, addr):
+        """Intercepts calls to :meth:`~slimta.edge.EdgeServer.handle`, reads
+        the proxy protocol header, and then resumes the original call.
+
+        """
+        try:
+            initial = self.__read_pp_initial(sock)
+            if initial.startswith(b'PROXY '):
+                src_addr, _ = ProxyProtocolV1.process_pp_v1(sock, initial)
+            elif initial == b'\r\n\r\n\x00\r\nQ':
+                src_addr, _ = ProxyProtocolV2.process_pp_v2(sock, initial)
+            else:
+                raise AssertionError('Invalid proxy protocol signature')
+        except LocalConnection:
+            log.proxyproto_local(sock)
+            return
+        except AssertionError as exc:
+            log.proxyproto_invalid(sock, exc)
+            src_addr = invalid_pp_source_address
+        else:
+            log.proxyproto_success(sock, src_addr)
+        return super(ProxyProtocol, self).handle(sock, src_addr)
 
 
 # vim:et:fdm=marker:sts=4:sw=4:ts=4