+++ test/test_slimta_util_proxyproto.py
@@ -2,7 +2,8 @@
 from mox3.mox import MoxTestBase, IsA
 from gevent import socket
 
-from slimta.util.proxyproto import ProxyProtocolV1
+from slimta.util.proxyproto import ProxyProtocol, \
+    ProxyProtocolV1, ProxyProtocolV2
 
 
 class SomeEdgeServer(object):
@@ -20,8 +21,69 @@ class SomeEdgeServer(object):
         self.handle_called = True
 
 
-class PPEdgeServer(ProxyProtocolV1, SomeEdgeServer):
+class PPV1Edge(ProxyProtocolV1, SomeEdgeServer):
     pass
+
+
+class PPV2Edge(ProxyProtocolV2, SomeEdgeServer):
+    pass
+
+
+class PPEdge(ProxyProtocol, SomeEdgeServer):
+    pass
+
+
+class TestProxyProtocol(MoxTestBase):
+    def setUp(self):
+        super(TestProxyProtocol, self).setUp()
+        self.pp = ProxyProtocol()
+
+    def test_read_pp_initial(self):
+        def _get_side_effect(data):
+            def _side_effect(view, length):
+                view[0:len(data)] = data
+            return _side_effect
+
+        sock = self.mox.CreateMock(socket.socket)
+        sock.recv_into(IsA(memoryview), 8).WithSideEffects(_get_side_effect(b'abcdefg')).AndReturn(7)
+        sock.recv_into(IsA(memoryview), 1).WithSideEffects(_get_side_effect(b'h')).AndReturn(1)
+        self.mox.ReplayAll()
+        line = self.pp._ProxyProtocol__read_pp_initial(sock)
+        self.assertEqual(b'abcdefgh', line)
+
+    def test_handle_pp_v1(self):
+        sock = self.mox.CreateMock(socket.socket)
+        sock.fileno = lambda: -1
+        pp = PPEdge(self, sock, 13)
+        self.mox.StubOutWithMock(ProxyProtocol, '_ProxyProtocol__read_pp_initial')
+        self.mox.StubOutWithMock(ProxyProtocolV1, 'process_pp_v1')
+        pp._ProxyProtocol__read_pp_initial(sock).AndReturn(b'PROXY ')
+        ProxyProtocolV1.process_pp_v1(sock, b'PROXY ').AndReturn((13, 14))
+        self.mox.ReplayAll()
+        pp.handle(sock, None)
+        self.assertTrue(pp.handle_called)
+
+    def test_handle_pp_v2(self):
+        sock = self.mox.CreateMock(socket.socket)
+        sock.fileno = lambda: -1
+        pp = PPEdge(self, sock, 13)
+        self.mox.StubOutWithMock(ProxyProtocol, '_ProxyProtocol__read_pp_initial')
+        self.mox.StubOutWithMock(ProxyProtocolV2, 'process_pp_v2')
+        pp._ProxyProtocol__read_pp_initial(sock).AndReturn(b'\r\n\r\n\x00\r\nQ')
+        ProxyProtocolV2.process_pp_v2(sock, b'\r\n\r\n\x00\r\nQ').AndReturn((13, 14))
+        self.mox.ReplayAll()
+        pp.handle(sock, None)
+        self.assertTrue(pp.handle_called)
+
+    def test_handle_error(self):
+        sock = self.mox.CreateMock(socket.socket)
+        sock.fileno = lambda: -1
+        pp = PPEdge(self, sock, (None, None))
+        self.mox.StubOutWithMock(ProxyProtocol, '_ProxyProtocol__read_pp_initial')
+        pp._ProxyProtocol__read_pp_initial(sock).AndRaise(AssertionError)
+        self.mox.ReplayAll()
+        pp.handle(sock, None)
+        self.assertTrue(pp.handle_called)
 
 
 class TestProxyProtocolV1(MoxTestBase):
@@ -66,9 +128,9 @@ class TestProxyProtocolV1(MoxTestBase):
     def test_handle(self):
         sock = self.mox.CreateMock(socket.socket)
         sock.fileno = lambda: -1
-        pp = PPEdgeServer(self, sock, 13)
-        self.mox.StubOutWithMock(pp, '_ProxyProtocolV1__read_pp_line')
-        self.mox.StubOutWithMock(pp, 'parse_pp_line')
+        pp = PPV1Edge(self, sock, 13)
+        self.mox.StubOutWithMock(ProxyProtocolV1, '_ProxyProtocolV1__read_pp_line')
+        self.mox.StubOutWithMock(ProxyProtocolV1, 'parse_pp_line')
         pp._ProxyProtocolV1__read_pp_line(sock, b'').AndReturn(b'the line')
         pp.parse_pp_line(b'the line').AndReturn((13, 14))
         self.mox.ReplayAll()
@@ -78,8 +140,8 @@ class TestProxyProtocolV1(MoxTestBase):
     def test_handle_error(self):
         sock = self.mox.CreateMock(socket.socket)
         sock.fileno = lambda: -1
-        pp = PPEdgeServer(self, sock, (None, None))
-        self.mox.StubOutWithMock(pp, '_ProxyProtocolV1__read_pp_line')
+        pp = PPV1Edge(self, sock, (None, None))
+        self.mox.StubOutWithMock(ProxyProtocolV1, '_ProxyProtocolV1__read_pp_line')
         pp._ProxyProtocolV1__read_pp_line(sock, b'').AndRaise(AssertionError)
         self.mox.ReplayAll()
         pp.handle(sock, None)
@@ -110,6 +172,82 @@ class TestProxyProtocolV1(MoxTestBase):
             f(b'PROXY TCP4 0.0.0.0 0.0.0.0 abc def\r\n')
         with self.assertRaises(AssertionError):
             f(b'PROXY TCP4 0.0.0.0 0.0.0.0 100000 0\r\n')
+
+
+class TestProxyProtocolV2(MoxTestBase):
+
+    def setUp(self):
+        super(TestProxyProtocolV2, self).setUp()
+        self.pp = ProxyProtocolV2()
+
+    def test_read_pp_data(self):
+        def _get_side_effect(data):
+            def _side_effect(view, length):
+                view[0:len(data)] = data
+            return _side_effect
+
+        sock = self.mox.CreateMock(socket.socket)
+        sock.recv_into(IsA(memoryview), 14).WithSideEffects(_get_side_effect(b'cdefghij')).AndReturn(8)
+        sock.recv_into(IsA(memoryview), 6).WithSideEffects(_get_side_effect(b'klmnop')).AndReturn(6)
+        self.mox.ReplayAll()
+        data = self.pp._ProxyProtocolV2__read_pp_data(sock, 16, b'ab')
+        self.assertEqual(b'abcdefghijklmnop', data)
+
+    def test_parse_pp_data(self):
+        data = bytearray(b'\r\n\r\n\x00\r\nQUIT\n\x21\x21\xf0\xf0')
+        cmd, fam, proto, addr_len = self.pp._ProxyProtocolV2__parse_pp_data(data)
+        self.assertEqual('proxy', cmd)
+        self.assertEqual(socket.AF_INET6, fam)
+        self.assertEqual(socket.SOCK_STREAM, proto)
+        self.assertEqual(61680, addr_len)
+
+    def test_parse_pp_addresses(self):
+        data = bytearray(b'\x00\x00\x00\x00\x7f\x00\x00\x01\x00\x00\x19\x00')
+        src_addr, dst_addr = self.pp._ProxyProtocolV2__parse_pp_addresses(socket.AF_INET, data)
+        self.assertEqual(('0.0.0.0', 0), src_addr)
+        self.assertEqual(('127.0.0.1', 25), dst_addr)
+        data = bytearray((b'\x00'*15 + b'\x01')*2 + b'\x00\x00\x19\x00')
+        src_addr, dst_addr = self.pp._ProxyProtocolV2__parse_pp_addresses(socket.AF_INET6, data)
+        self.assertEqual(('::1', 0), src_addr)
+        self.assertEqual(('::1', 25), dst_addr)
+        data = bytearray(b'abc' + b'\x00'*105 + b'def' + b'\x00'*105)
+        src_addr, dst_addr = self.pp._ProxyProtocolV2__parse_pp_addresses(socket.AF_UNIX, data)
+        self.assertEqual(b'abc', src_addr)
+        self.assertEqual(b'def', dst_addr)
+
+    def test_process_pp_v2(self):
+        sock = self.mox.CreateMock(socket.socket)
+        self.mox.StubOutWithMock(ProxyProtocolV2, '_ProxyProtocolV2__read_pp_data')
+        self.mox.StubOutWithMock(ProxyProtocolV2, '_ProxyProtocolV2__parse_pp_data')
+        self.mox.StubOutWithMock(ProxyProtocolV2, '_ProxyProtocolV2__parse_pp_addresses')
+        ProxyProtocolV2._ProxyProtocolV2__read_pp_data(sock, 16, b'init').AndReturn(b'data1')
+        ProxyProtocolV2._ProxyProtocolV2__parse_pp_data(b'data1').AndReturn((1, 2, 3, 4))
+        ProxyProtocolV2._ProxyProtocolV2__read_pp_data(sock, 4, b'').AndReturn(b'data2')
+        ProxyProtocolV2._ProxyProtocolV2__parse_pp_addresses(2, b'data2').AndReturn((13, 14))
+        self.mox.ReplayAll()
+        src_addr, dst_addr = ProxyProtocolV2.process_pp_v2(sock, b'init')
+        self.assertEqual(13, src_addr)
+        self.assertEqual(14, dst_addr)
+
+    def test_handle(self):
+        sock = self.mox.CreateMock(socket.socket)
+        sock.fileno = lambda: -1
+        pp = PPV2Edge(self, sock, 13)
+        self.mox.StubOutWithMock(ProxyProtocolV2, 'process_pp_v2')
+        ProxyProtocolV2.process_pp_v2(sock, b'').AndReturn((13, 14))
+        self.mox.ReplayAll()
+        pp.handle(sock, None)
+        self.assertTrue(pp.handle_called)
+
+    def test_handle_error(self):
+        sock = self.mox.CreateMock(socket.socket)
+        sock.fileno = lambda: -1
+        pp = PPV2Edge(self, sock, (None, None))
+        self.mox.StubOutWithMock(ProxyProtocolV2, 'process_pp_v2')
+        ProxyProtocolV2.process_pp_v2(sock, b'').AndRaise(AssertionError)
+        self.mox.ReplayAll()
+        pp.handle(sock, None)
+        self.assertTrue(pp.handle_called)
 
 
 # vim:et:fdm=marker:sts=4:sw=4:ts=4